Refactoring in order to work with new services
