+++ public/controllers/reporting.js
@@ -9,8 +9,10 @@
  *
  * Find more information about this on the LICENSE file.
  */
-const app = require('ui/modules').get('app/wazuh', []);
+import * as modules   from 'ui/modules'
 import $ from 'jquery'
+const app = modules.get('app/wazuh', []);
+
 // Logs controller
 app.controller('reportingController', function ($scope, errorHandler, genericReq) {
     $scope.loading = true;
